// src/lib/locations.generated.ts
export type GeneratedLocation = {
  slug: string
  city: string
  state: string
  stateCode: string
  county: string
}